Test: split User from models.py
