+++ app/event/helpers.py
@@ -1,5 +1,6 @@
 from app import db
-from app.models import Event, EventCategory, Role, User
+from app.models import Event, EventCategory
+from app.user.models import Role, User
 from app.calendar.helpers import gen_calendar_stats
 from flask import redirect, url_for, flash
 from functools import wraps
